fixed bug in float types

+++ pkg/job/job.go
@@ -41,14 +41,14 @@ type AdapterConfig map[string]QueryConfig // map of query string to its config
 
 // QueryConfig represents the configuration for a specific query
 type QueryConfig struct {
-	Pos      int  `json:"pos"`
-	QLen     int  `json:"qlen"`
-	Limit    int  `json:"limit"`
-	Tasks    int  `json:"tasks"`
-	Active   bool `json:"active"`
-	Enabled  bool `json:"enabled"`
-	Timeout  int  `json:"timeout"`
-	AutoTask bool `json:"autotask"`
+	Pos      int     `json:"pos"`
+	QLen     int     `json:"qlen"`
+	Limit    int     `json:"limit"`
+	Tasks    int     `json:"tasks"`
+	Active   bool    `json:"active"`
+	Enabled  bool    `json:"enabled"`
+	Timeout  float64 `json:"timeout"`
+	AutoTask bool    `json:"autotask"`
 }
 
 // set default values and validation here
